remove unnecessary Services class
