+++ selenium/che-selenium-test/src/test/java/org/eclipse/che/selenium/refactor/methods/RenameStaticMethodsTest.java
@@ -20,6 +20,8 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.Random;
+
+import org.eclipse.che.commons.lang.NameGenerator;
 import org.eclipse.che.selenium.core.client.TestProjectServiceClient;
 import org.eclipse.che.selenium.core.project.ProjectTemplates;
 import org.eclipse.che.selenium.core.workspace.TestWorkspace;
@@ -31,7 +33,6 @@ import org.eclipse.che.selenium.pageobject.Loader;
 import org.eclipse.che.selenium.pageobject.NotificationsPopupPanel;
 import org.eclipse.che.selenium.pageobject.ProjectExplorer;
 import org.eclipse.che.selenium.pageobject.Refactor;
-import org.eclipse.che.selenium.refactor.Services;
 import org.openqa.selenium.Keys;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -42,12 +43,11 @@ import org.testng.annotations.Test;
 
 /** @author Musienko Maxim */
 public class RenameStaticMethodsTest {
-  private static final Logger LOG = LoggerFactory.getLogger(RenameStaticMethodsTest.class);
-  private static final String nameOfProject =
-      RenameStaticMethodsTest.class.getSimpleName() + new Random().nextInt(9999);
+  private static final Logger LOG                      = LoggerFactory.getLogger(RenameStaticMethodsTest.class);
+  private static final String NAME_OFP_ROJECT          = NameGenerator.generate(RenameStaticMethodsTest.class.getSimpleName(), 3);
   private static final String pathToPackageInChePrefix =
-      nameOfProject + "/src" + "/main" + "/java" + "/renameStaticMethods";
-  private static final String testsFail5ErrorMess =
+          NAME_OFP_ROJECT + "/src" + "/main" + "/java" + "/renameStaticMethods";
+  private static final String testsFail5ErrorMess      =
       "Related method 'm' (declared in 'renameStaticMethods.testFail5.A') is native. Renaming will cause an UnsatisfiedLinkError on runtime.";
 
   private String pathToCurrentPackage;
@@ -74,11 +74,11 @@ public class RenameStaticMethodsTest {
     testProjectServiceClient.importProject(
         workspace.getId(),
         Paths.get(resource.toURI()),
-        nameOfProject,
+        NAME_OFP_ROJECT,
         ProjectTemplates.MAVEN_SIMPLE);
     ide.open(workspace);
-    new Services(projectExplorer, notificationsPopupPanel, refactor)
-        .expandRenamePrivateMethodProject(nameOfProject, "renameStaticMethods");
+    projectExplorer.waitItem(NAME_OFP_ROJECT);
+    projectExplorer.quickExpandWithJavaScript();
     consoles.closeProcessesArea();
   }
 
@@ -180,7 +180,6 @@ public class RenameStaticMethodsTest {
 
   private void prepareProjectForRefactor(int cursorPositionLine, int cursorPositionChar) {
     projectExplorer.waitItem(pathToPackageInChePrefix);
-    projectExplorer.openItemByPath(pathToCurrentPackage);
     projectExplorer.openItemByPath(pathToCurrentPackage + "/A.java");
     editor.waitTextIntoEditor(contentFromInA);
     editor.goToCursorPositionVisible(cursorPositionLine, cursorPositionChar);
